Общее: ~ Улучшение опции «Строгие имена файлов» - теперь недопустимые кавычки “ в названиях fb2 файла заменяются на допустимые «елочки «» ». Дубликатор: + Новый элемент контекстного меню и панели инструментов: Правка Названия Книги. Сортировщик: ~ Улучшение работы возобновления сортировки из xml-файла, если используется шаблон подстановки *COUNTER*.

+++ Source/Core/FileManager/StatusView.cs
@@ -16,18 +16,18 @@ namespace Core.FileManager
 	public class StatusView {
 		
 		#region Закрытые данные класса
-		private int m_nAllDirs			= 0;
-		private int m_nAllFiles			= 0;
-		private int m_nSourceFB2		= 0;
-		private int m_nZip				= 0;
-		private int m_nFB2FromZips		= 0;
-		private int m_nOther			= 0;
-		private int m_nCreateInTarget	= 0;
-		private int m_nNotRead			= 0;
-		private int m_nNotValidFB2		= 0;
-		private int m_nBadZip			= 0;
-		private int m_nNotSort			= 0;
-		private int m_nLongPath			= 0;
+		private long m_nAllDirs			= 0;
+		private long m_nAllFiles		= 0;
+		private long m_nSourceFB2		= 0;
+		private long m_nZip				= 0;
+		private long m_nFB2FromZips		= 0;
+		private long m_nOther			= 0;
+		private long m_nCreateInTarget	= 0;
+		private long m_nNotRead			= 0;
+		private long m_nNotValidFB2		= 0;
+		private long m_nBadZip			= 0;
+		private long m_nNotSort			= 0;
+		private long m_nLongPath		= 0;
 		#endregion
 		
 		public StatusView() {
@@ -53,51 +53,51 @@ namespace Core.FileManager
 		#endregion
 		
 		#region Свойства класса
-		public virtual int AllDirs {
+		public virtual long AllDirs {
 			get { return m_nAllDirs; }
 			set { m_nAllDirs = value; }
         }
-		public virtual int AllFiles {
+		public virtual long AllFiles {
 			get { return m_nAllFiles; }
 			set { m_nAllFiles = value; }
         }
-		public virtual int SourceFB2 {
+		public virtual long SourceFB2 {
 			get { return m_nSourceFB2; }
 			set { m_nSourceFB2 = value; }
         }
-		public virtual int Zip {
+		public virtual long Zip {
 			get { return m_nZip; }
 			set { m_nZip = value; }
         }
-		public virtual int FB2FromZips {
+		public virtual long FB2FromZips {
 			get { return m_nFB2FromZips; }
 			set { m_nFB2FromZips = value; }
         }
-		public virtual int Other {
+		public virtual long Other {
 			get { return m_nOther; }
 			set { m_nOther = value; }
         }
-		public virtual int CreateInTarget {
+		public virtual long CreateInTarget {
 			get { return m_nCreateInTarget; }
 			set { m_nCreateInTarget = value; }
         }
-		public virtual int NotRead {
+		public virtual long NotRead {
 			get { return m_nNotRead; }
 			set { m_nNotRead = value; }
         }
-		public virtual int NotValidFB2 {
+		public virtual long NotValidFB2 {
 			get { return m_nNotValidFB2; }
 			set { m_nNotValidFB2 = value; }
         }
-		public virtual int BadZip {
+		public virtual long BadZip {
 			get { return m_nBadZip; }
 			set { m_nBadZip = value; }
         }
-		public virtual int NotSort {
+		public virtual long NotSort {
 			get { return m_nNotSort; }
 			set { m_nNotSort = value; }
 		}
-		public virtual int LongPath {
+		public virtual long LongPath {
 			get { return m_nLongPath; }
 			set { m_nLongPath = value; }
 		}
